Use cssnano through postcss in gulp-tasks
The previously used package `gulp-cssnano` is deprecated and uses
an outdated version of Cssnano. Requiring cssnano directly and running
it through postcss allows for updating it to the newest version, as well
as better performance (because of the reused AST).

Also added was the option to configure cssnano by using a config object:

```
gulpConfig.files.bundle.cssnanoSettings = {
  discardComments: {
    removeAll: true
  }
};
```