fix: improve obtaining currently routed path
When routing to unknown path, we need to pass that route (where we want to go) within the action object further. Later we can check, whether that path exists in plugin-redirect. getUrl method was returning the current path (where we are). We also need to remove error condition, so the set path is not thrown away. The condition was there most likely due to devtools and is not needed anymore.